added running version of pyncp
when running pyncp as main it will either run in listen mode
when no arguments are given or in push mode when the next arg is a
hostname/ip and ther further args are filenames.

pyncpmain might be the wrapper for the ncp and later on the push and
poll functionality ( similar to the original ncp ). anyway i am not
yet sure about that

updated gitignore to ignore .pyc files